initial

+++ src/main/java/com/anteater/star/domain/Bill.java
@@ -2,12 +2,13 @@ package com.anteater.star.domain;
 
 import lombok.Getter;
 import lombok.Setter;
+import org.springframework.stereotype.Component;
 
 import java.util.Date;
 
 @Getter@Setter
 public class Bill {
-    private int billId;
+    private Long billId;
     private String billYear;
     private String billMonth;
     private String billDate;
@@ -19,10 +20,11 @@ public class Bill {
     private Double billRubbish;
     private Double billOther;
     //Check the bill is pay or not.
-    private int billIsPay;
+    private Long billIsPay;
     private Date billPayDate;
     private String billPayWay;
     //fk
     private String billRoomId;
+    
     private Room billRoom;
 }
