+++ src/main/java/com/anteater/star/domain/Room.java
@@ -2,25 +2,29 @@ package com.anteater.star.domain;
 
 import lombok.Getter;
 import lombok.Setter;
+import org.springframework.stereotype.Component;
 
 import java.util.Set;
 
 @Getter@Setter
 public class Room {
-    private int roomId;
+    private Long roomId;
     private String roomName;
     //单间或套间
     private String roomType;
-    private int roomFloor;
-    private int roomStatus;
+    private Integer roomFloor;
+    private Integer roomStatus;
     private String roomContract;
     //是否具有空调
     private String roomFacility;
     //fk
     //house
-    private int roomHouseId;
+    private Long roomHouseId;
+
+    //many to one object
     private House roomHouse;
-    //bill
+
+    //bill many to many
     private Set<Bill> roomBillSet;
     //Renter
     private Set<Renter> roomRenters;
